# System_BE/Swagger/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Smart Amenities Backend API",
    "description": "Backend API\nReference https://swagger-autogen.github.io/docs/ for details on configuring the swagger output. ",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": [
    "http",
    "ws"
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/nav/map": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "location",
            "in": "body",
            "description": "send the location of the user with zoom level to center the map",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/locationWithZoom"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/nav": {
      "get": {
        "description": "This endpoint IS NOT an HTTP GET request endpoint - it is used with websockets protocol. Use this endpoint to create a websocket conection. ",
        "responses": {
          "200": {
            "description": "Creating websocket connection"
          }
        }
      }
    },
    "/amenities/all": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "location",
            "in": "body",
            "description": "send the location of the user",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/location"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/amenities/oftype": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "location",
            "in": "body",
            "description": "send the location of the user",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/location"
            }
          },
          {
            "name": "amenityType",
            "in": "body",
            "description": "send the type of amnenity desired",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/amenityType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/amenities/suggested": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "location",
            "in": "body",
            "description": "send the location of the user",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/location"
            }
          },
          {
            "name": "filters",
            "in": "body",
            "description": "send the filters to apply",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/filters"
            }
          },
          {
            "name": "sortMethod",
            "in": "body",
            "description": "send the sort method to use",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/sortMethod"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/amenities/details": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "amenityID",
            "in": "body",
            "description": "send the id of the desired amenity",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/amenityID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/amenities/filter": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "filters",
            "in": "body",
            "description": "send the filters to apply",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/filters"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/amenities/updateoccupancy": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "body",
            "description": "id oft eh amenity whose occupancy is being updated",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/amenityID"
            }
          },
          {
            "name": "details",
            "in": "body",
            "description": "new occupancy data for the amenity",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/occupancyDetails"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "502": {
            "description": "Bad Gateway"
          }
        }
      }
    },
    "/amenities/notification": {
      "get": {
        "description": "This endpoint is called when the amenities are updated",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "location": {
        "type": "object",
        "properties": {
          "x": {
            "type": "string"
          },
          "y": {
            "type": "string"
          }
        },
        "required": [
          "x",
          "y"
        ],
        "example": {
          "x": "10.11",
          "y": "12.13"
        }
      },
      "sortMethod": {
        "type": "object",
        "properties": {
          "sortMethod": {
            "type": "string"
          }
        },
        "required": [
          "sortMethod"
        ],
        "example": {
          "sortMethod": "BEST_ROUTE"
        }
      },
      "locationWithZoom": {
        "type": "object",
        "properties": {
          "x": {
            "type": "string"
          },
          "y": {
            "type": "string"
          },
          "zoom": {
            "type": "number"
          }
        },
        "required": [
          "x",
          "y",
          "zoom"
        ],
        "example": {
          "x": "10.11",
          "y": "12.13",
          "zoom": 15
        }
      },
      "amenityID": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "example": {
          "id": "AM2"
        }
      },
      "filters": {
        "type": "object",
        "properties": {
          "filters": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "filterKey": {
                  "type": "string"
                },
                "value": {
                  "type": "string"
                }
              },
              "required": [
                "filterKey",
                "value"
              ],
              "example": {
                "filterKey": "type",
                "value": "RESTROOM"
              }
            }
          }
        },
        "required": [
          "filters"
        ],
        "example": [
          {
            "filterKey": "type",
            "value": "RESTROOM"
          }
        ]
      },
      "amenityType": {
        "type": "object",
        "properties": {
          "amenityType": {
            "type": "string"
          }
        },
        "required": [
          "amenityType"
        ],
        "example": {
          "amenityType": "RESTROOM"
        }
      },
      "occupancyDetails": {
        "type": "object",
        "properties": {
          "currentOccupancy": {
            "type": "number"
          },
          "currentAvailableSlots": {
            "type": "number"
          },
          "capacity": {
            "type": "number"
          },
          "status": {
            "type": "string"
          },
          "lastUpdated": {
            "type": "string"
          }
        },
        "required": [
          "currentOccupancy",
          "currentAvailableSlots",
          "capacity",
          "status",
          "lastUpdated"
        ],
        "example": {
          "currentOccupancy": 2,
          "currentAvailableSlots": 4,
          "capacity": 6,
          "status": "OPEN",
          "lastUpdated": "03/29/26, 10:10:31"
        }
      }
    }
  }
}